修正

+++ 舰船服务器/item_pt.c
@@ -710,9 +710,9 @@ int pt_bb_enabled(void) {
 	return have_bbpt;
 }
 
-int get_pt_index(uint8_t episode, uint8_t pt_index) {
-	uint8_t ep4_pt_index_offset = 0x57;//87 Item_PT EP4 enemy_index ²îÖµ
-	return episode == GAME_TYPE_EPISODE_3 ? pt_index - ep4_pt_index_offset : episode == GAME_TYPE_EPISODE_4 ? pt_index - ep4_pt_index_offset : pt_index;
+size_t get_pt_index(uint8_t episode, size_t pt_index) {
+	size_t ep4_pt_index_offset = 0x57;//87 Item_PT EP4 enemy_index ²îÖµ
+	return (episode == GAME_TYPE_EPISODE_3 ? (pt_index - ep4_pt_index_offset) : episode == GAME_TYPE_EPISODE_4 ? (pt_index - ep4_pt_index_offset) : pt_index);
 }
 
 int get_pt_data_area_bb(uint8_t episode, int cur_area) {
@@ -3828,7 +3828,7 @@ int pt_generate_bb_drop(ship_client_t* src, lobby_t* l, void* r) {
 	game_enemy_t* enemy;
 	int csr = 0;
 	uint8_t pt_index = req->rt_index;
-	int ep_pt_index = get_pt_index(l->episode, pt_index);
+	size_t ep_pt_index = get_pt_index(l->episode, pt_index);
 
 	pt_bb_entry_t* ent = get_pt_data_bb(l->episode, l->challenge, l->difficulty, section);
 	if (!ent) {
@@ -4304,7 +4304,7 @@ int pt_generate_bb_pso2_drop_style(ship_client_t* src, lobby_t* l, uint8_t secti
 	game_enemy_t* enemy;
 	int csr = 0;
 	uint8_t pt_index = req->rt_index;
-	int ep_pt_index = get_pt_index(l->episode, pt_index);
+	size_t ep_pt_index = get_pt_index(l->episode, pt_index);
 
 	pt_bb_entry_t* ent = get_pt_data_bb(l->episode, l->challenge, l->difficulty, section);
 	if (!ent) {
